[WIP] Download released 'flatc'
Instead of building it from source, download the released 'flatc' binary from GitHub.

For the binary to work in the currently used build image, packages from the "testing" branch have to be installed due to the recent version of the C++ runtime that the program was compiled with.

The current form of this change is therefore not considered mergeable. This might change in the future. In any case, it's useful to keep it as documentation of why this isn't doable for now.