Docker Compose: Configure init containers
Using the new "condition" feature of the 'depends_on' field to properly defer the creation of the 'node' container until the init containers 'node-init-genesis' and 'node-init-catchup' have completed executing.

This replaces the shell-based logic that was used to implement the behavior before.

The entrypoint command of the genesis image was also moved to the Compose script (and statefulset in the Helm chart) as part of this.