Docker Compose: Don't re-fetch OOB data
If the OOB blocks file is already present, then we assume that we don't need to do a "full" enough catchup to warrant a download of such a large file.

However, there may be cases where the node did originally boot using the file, but since then, so much time has passed that doing OOB cachup would actually be worthwhile. An improvement would thus be to still re-fetch if the file is older than a certain age, although it isn't clear what such a limit should be.

An option to force refetching could also be added. Until then, a manual way to achieve this is to delete the file from the volume using a throwaway container; i.e.

    docker run --rm -v concordium-docker_data:/data busybox:latest rm /data/blocks.mdb