OOB catchup: Use 'curl' instead of 'wget'
Busybox' variant of Wget doesn't support certificate validation, but curl does.

The used image is officially supported by the curl team. As this image uses user 'curl_user' (UID 100) by default, we explicitly change it to 'root' to match the other containers and give it permission to write to the shared storage.