Genesis: Use bind mount instead of init container
The dockerfile for building the genesis image has been kept, but its usage removed from the Docker Compose script. Likewise, it's no longer being built in the CI pipeline. Instead, the genesis file is supposed to be injected via a bind mount. The current testnet genesis has been added as part of this.

Additionally, in the Docker Compose setup, all volumes are now mounted inside '/mnt'.