Replace domain prefix with suggestion to set project name
Unless the 'container_name' is explicitly set, the container name will default to the service name prefixed by any given project name.

Compared to prepending the domain (which in e39e320a was believed to be the only working solution), this solution makes the log output cleaner by not including the prefix. It also allows one to use shorter names than the somewhat verbose domain name.