Compose: Stop collector using SIGKILL
Deployments currently take a long time to stop, which turns out to be because the collector ignores SIGINT and SIGTERM.

The node actually turns out to stop very quickly after receiving SIGTERM, but it blocked by the collector because of the dependency relationship.

Nevertheless, a note on increasing the stop timeout in case the node needs extra time has been added for good measure. This resolves #14.