Compose: Add ability to re-fetch OOB data (#17)
The value of the environment variable 'OOB_CATCHUP_REFRESH_AGE_SECS' is compared against the modification time of 'lock.mdb'. If it's older, a re-fetch is triggered. The refresh age defaults to never, but is set to 14 days in the env files.

The node actually does reengage OOB catchup with the updated file, though it isn't clear what (if any) performance benefit it has compared to standard catchup: In the test it took about 10 mins before it started processing blocks, and afterwards did fully catch up within a reasonable amount of time. The delay is probably caused by the whole file having to be processed.

To determine the optimal refresh age value, benchmarks would have to be set up.

Resolves #9.