Merge env files for running with or without transaction logging
Since transaction logging has been migrated to a separate service, there's no longer any need for running the two modes in different projects - whenever the profile is enabled, the logger will simply catch up from where it left last time it was run. It doesn't matter if the node was running without the logger in the meantime.

As a result, the "txlog" variables are now always set, but only have any effect when the 'txlog' profile is enabled.

The variable 'COMPOSE_PROFILE' has also been removed from the env files. Instead, the run script now supports adding profiles as extra arguments of the form "+<profile>".