Compose: Move OOB to "override" compose file
A deployment may be created from multiple Compose files, which are merged in the order that they're listed. This is used to extract the OOB service and node dependency into its own file which is activated as a unit - something which annoyingly isn't possible with profiles.

The 'run' script is updated to interpret "+<feature>" as both enabling profiles and loading "docker-compose.<feature>.yaml" if it exists.