Compose: Move OOB feature to "override" compose file (#51)
A deployment may be created from multiple Compose files, which are merged in the order that they're listed. This is used to extract the OOB service and node dependency into its own file which is activated as a unit - something which annoyingly isn't possible with profiles.

The 'run.sh' script is updated to interpret '+<feature>' as both enabling profiles and loading 'docker-compose.<feature>.yaml' if it exists. This will allow other features to be implemented as overrides in the future.

Features that might make sense to also do as overrides:

* Metrics (because it adds a volume).
* Transaction logger: Add ability to use PostgreSQL in the deployment instead of requiring an external one to be available.