Compose: Add CCDScan to deployment as an override (#53)
Adding feature ccdscan as an override (because it adds a new volume for the DB).

In a followup change, the front- and backends should be moved to the main Compose spec. Only the TimescaleDB running inside the deployment should be in an override. There should also be a new override for running the transaction logger PostgreSQL DB inside the deployment. These changes might as well be done together.