Compose: Put OOB data in dedicated volume (#52)
This separates the OOB file which gets stale immediately from the DB which stays in active use, adding confidence that the file (or the whole volume for that matter) may be wiped without unintended side effects. It also removes the need for an '--exclude' exception in the command for backing up data.