Add reference to commit that removed legacy OOB
Description missing from previous commit:

The node no longer requires a file to be downloaded in advance in order to use OOB catchup. It now accepts a URL from which it can figure out what chunks it needs and will download them by itself. The old mode is still available, but the exported file is no longer being updated. There is therefore little reason to keep supporting it.