Add tool for resolving catchup status for a given node
The tool currently just prints the block height of the chain relative to the provided node. The node is provided by name and the chain's current height is determined by a given percentile of the reported heights for the nodes that are deemed up to date. The nodes that are deemed up to date is again determined as a percentile of the reporting nodes' reported versions.

There are a number of possible improvements to be made mentioned in TODOs in the file.