Add build and deploy config for the old Network Dashboard
Currently only the "nodes" page works (i.e. no chain explorer). It's unclear if this can be made to work with the current version of concordium-node as it no longer supports gRPC APIv1 and I'm unsure if "middleware" was ever migrated to v2. If not, it may be too much effort to get this part running. But the plan is to give it a try.

Usage is documented in the new Compose file, but not yet in the readme. This will (hopefully) be done later.