Network Dashboard: Make container stop on SIGTERM
The Docker Compose file specifies a shell script for the container entrypoint. This is necessary (at least as long as we don't use templates; see comment in `docker-compose.yaml` right above the changed line) to set up redirection for the calls to Collector Backend.

The script ends with a call to the native nginx entrypoint script. But apparently, the script running in the context of a plain shell means that SIGTERM is ignored. With this change, the entrypoint script is now invoked with `exec` which replaces the process and eliminates the problem.