Compose: Migrate to Docker Compose v2
Docker Compose v1 CLI (written in Python) has been sunset for a while now in favor of v2, which is implemented as a plugin to the `docker` command and written in Go. Fortunately, the implementations are sufficiently compatible that migration is a simple matter of invoking the command as `docker compose` instead of `docker-compose`.

So the change simple removes the old Python dependency and performs the aforementioned string replacement.

It might be worth looking into whether some newer features of v2 could be used to simplify the deployment, but as of this commit this hasn't been done and the project therefore remains entirely compatible with v1 for now.