CI: Add placeholders for unused variables
Placeholder/dummy values for variables that are either not used by the build command or belong to services that aren't active have been added to the build workflow.

This is done to satisfy a new requirement in Compose v2 that wasn't caught in 3b44b2a0f72f2e3e363e831a4e238a3deb28cdd7.