Docker: Bump GHC version to fix build for Node 6.3.0
The Node has been bumped from using Stackage LTS-20.16 (GHC 9.2.7) to LTS-22.9 (GHC 0.6.4). This requires us to bump the Haskell build image accordingly to fix the build.

The default value of the 'tag' describes (as documented) the oldest version of the node that the build file has been verified to work with and has been bumped accordingly.

Note that the new Node version hasn't been released yet, so we don't bump "active tag" in the README.