Docker: Add missing shared library for consensus
This is a new library as of Node 6.3.0. It built fine without it but failed on startup with a nice and clear error message:

	/concordium-node: error while loading shared libraries: libconcordium-consensus.so: cannot open shared object file: No such file or directory

With this change the node works again.

I don't know why the library is now there or whether it's replacing some of the other include paths; at some point we ought to figure out which ones are actually used.