run.sh: Remove pull command
The [Compose spec](https://github.com/compose-spec/compose-spec/blob/master/build.md#build-vs-image) states that with the configuration we're using, Compose will attempt to pull our images before building them locally:

> When Compose is confronted with both a `build` subsection for a service and an `image` attribute. It follows the rules defined by the [`pull_policy`](https://github.com/compose-spec/compose-spec/blob/master/05-services.md#pull_policy) attribute.
>
> If `pull_policy` is missing in the service definition, Compose attempts to pull the image first and then builds from source if the image isn't found in the registry or platform cache.

This is the behavior corrently observed with Compose v2.20.2, so the pull command is unnecessary and in fact breaks `ccdscan_frontend` because its image doesn't exist in the registry (it's supposed to be built locally), so attempting to pull it will fail.

Digging back into the spec it looks like that was always the intended behavior, at least back to the time when v1.29.2 (the last v1) was released. But trying with v1.29.2 again confirms that it did indeed go ahead and build without trying to pull first.