catchup-status: Lock dependencies to minor verion
It looks like I've previously misunderstood the "compatible version" specifier, but as documented in https://peps.python.org/pep-0440/#compatible-release, the expression `~= x.y` actually means "compatible up to the next major version".

Adding a `.0` locks to compatibility with the given minor version instead.

This change is necessary because the script appears to be incompatible with pydantic 2.6.x (it was written for 2.3.x). With this change, we're reverting to use the original version.