Update Dockerfile for Node tag 8.0.3-1
This required a bump to GHC (and thus Debian base) and Rust versions.

The docs are updated appropriately. In particular, the inclusion of the currently active tag is removed from the README; it's not useful there with the repo no longer being updated regularly.

The .env files have not been updated as no builds have been published yet.